tunefs.ocfs2: Use uint32_t instead of uint64_t for total_clusters.
In ocfs2, we always use uint32_t to indicate cluster count.
As total_clusters is always used as a uint32_t below, it must
be defined as uint64_t by mistake. So change it as uint32_t.

+++ tunefs.ocfs2/sparse_file.c
@@ -36,7 +36,7 @@ struct multi_link_file {
 
 struct list_ctxt {
 	ocfs2_filesys *fs;
-	uint64_t total_clusters;
+	uint32_t total_clusters;
 	char file_name[OCFS2_MAX_FILENAME_LEN];
 	int file_name_len;
 	uint64_t ino;
